+++ vite.config.js
@@ -1,10 +1,9 @@
-const { defineConfig } = require('vite');
-const virtualIndex = require("@ui5/webcomponents-tools/lib/dev-server/virtual-index-html-plugin.js");
-const customHotUpdate = require("@ui5/webcomponents-tools/lib/dev-server/custom-hot-update-plugin.js");
-const { dirname, join, resolve } = require('path');
-const path = require('path');
-const tsconfigPaths = require('vite-tsconfig-paths').default;
-const { checker } = require('vite-plugin-checker');
+import { defineConfig } from 'vite';
+import virtualIndex from "@ui5/webcomponents-tools/lib/dev-server/virtual-index-html-plugin.js";
+import customHotUpdate from "@ui5/webcomponents-tools/lib/dev-server/custom-hot-update-plugin.js";
+import path, { dirname, join, resolve } from 'path';
+import tsconfigPaths from 'vite-tsconfig-paths';
+import { checker } from 'vite-plugin-checker';
 
 // use after path.join and path.resolve as they turn paths to windows separators and comparisons and replacements stop working
 const toPosixPath = (pathStr) => {
@@ -94,7 +93,7 @@ const customResolver = (id, source, options) => {
 	}
 }
 
-module.exports = defineConfig(async () => {
+export default defineConfig(async () => {
 	return {
 		build: {
 			emptyOutDir: false,